pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "velib-ingestion"
version = "0.1.0"
description = "Client and transformers for Velib Metropole GBFS open data"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31",
  "python-dotenv>=1.0",
  "pandas>=2.2",
  "tenacity>=8.2",
]

[project.optional-dependencies]
parquet = ["pyarrow>=14.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
